requirements.txt
streamlit
pandas
plotly
openpyxl
fpdf2